+++ src/providers/wixMarketing.js
@@ -1,13 +1,35 @@
 import { createCompletionList, K } from './util';
 
 const list = createCompletionList([
-  { name: 'coupons', kind: K.Property },
+  {
+    name: 'coupons',
+    kind: K.Property,
+    docs: 'The Coupons API is used to manage your site\'s coupons.',
+  },
 ]);
 
 const listCoupons = createCompletionList([
-  { name: 'createCoupon', kind: K.Method },
-  { name: 'deleteCoupon', kind: K.Method },
-  { name: 'updateCouponFields', kind: K.Method },
+  {
+    name: 'createCoupon',
+    kind: K.Method,
+    snippet: 'createCoupon(${1:couponInfo})',
+    detail: 'function createCoupon(couponInfo: CouponInfo): Promise<string>',
+    docs: 'Creates a new coupon.',
+  },
+  {
+    name: 'deleteCoupon',
+    kind: K.Method,
+    snippet: 'deleteCoupon(${1:couponId})',
+    detail: 'function deleteCoupon(couponId: string): Promise<void>',
+    docs: 'Deletes an existing coupon.',
+  },
+  {
+    name: 'updateCouponFields',
+    kind: K.Method,
+    snippet: 'updateCouponFields(${1:couponId}, ${2:couponData})',
+    detail: 'function updateCouponFields(couponId: string, couponInfo: CouponInfo): Promise<void>',
+    docs: 'Updates the specified fields of an existing coupon.',
+  },
 ]);
 
 export default {