+++ src/providers/index.js
@@ -1,7 +1,10 @@
-import { languages } from 'vscode';
+import { languages, workspace } from 'vscode';
 
 import modules from './modules';
 import roles from './roles';
+
+const config = workspace.getConfiguration('corvid.autocomplete', null);
+const providers = [];
 
 function register(provider, trigger = '.') {
   return languages.registerCompletionItemProvider(
@@ -11,7 +14,12 @@ function register(provider, trigger = '.') {
   );
 }
 
-export default [
-  register(modules, '.'),
-  register(roles, '#'),
-];
+if (config.get('import')) {
+  providers.push(register(modules, '.'));
+}
+
+if (config.get('$w')) {
+  providers.push(register(roles, '#'));
+}
+
+export default providers;