+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## v1.1.0
+* Added extension configuration
+
 ## v1.0.0
 * Autocomplete for selectors `$w()` (for Corvid projects)
 * Validation `jobs.config` file.